# openAPI-def.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "doughleap",
    "version": "1.0.0"
  },
  "paths": {
...
    "/steps": {
      "get": {
        "tags": [
          "Recipe Steps"
        ],
        "operationId": "getSteps",
        "parameters": [
          {
            "name": "stepIDs",
            "in": "query",
            "description": "Comma-separated list of step IDs.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "title",
            "in": "query",
            "description": "The title of the step.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "stepID": {
                        "type": "integer",
                        "description": "The ID of the step."
                      },
                      "title": {
                        "type": "string",
                        "description": "The title of the step."
                      },
                      "description": {
                        "type": "string",
                        "description": "The description of the step."
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recipe Steps"
        ],
        "operationId": "createStep",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The title of the step."
                  },
                  "description": {
                    "type": "string",
                    "description": "The description of the step."
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stepID": {
                      "type": "integer",
                      "description": "The ID of the step."
                    },
                    "title": {
                      "type": "string",
                      "description": "The title of the step."
                    },
                    "description": {
                      "type": "string",
                      "description": "The description of the step."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/steps/{stepID}": {
      "patch": {
        "tags": [
          "Recipe Steps"
        ],
        "operationId": "updateStep",
        "parameters": [
          {
            "name": "stepID",
            "in": "path",
            "description": "The ID of the step.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "The updated title of the step."
                  },
                  "description": {
                    "type": "string",
                    "description": "The updated description of the step."
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stepID": {
                      "type": "integer",
                      "description": "The ID of the step."
                    },
                    "title": {
                      "type": "string",
                      "description": "The title of the step."
                    },
                    "description": {
                      "type": "string",
                      "description": "The description of the step."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Recipe Steps"
        ],
        "operationId": "deleteStep",
        "parameters": [
          {
            "name": "stepID",
            "in": "path",
            "description": "The ID of the step.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
...
}